Changed to no parallel predict

+++ src/adaXT/decision_tree/nodes.pyx
@@ -30,6 +30,7 @@ cdef class DecisionNode(Node):
             is_leaf: int = 0,
             visited: int = 0) -> None:
 
+
         super().__init__(indices, depth, impurity)
         self.threshold = threshold
         self.split_idx = split_idx
@@ -51,7 +52,7 @@ cdef class LeafNode(Node):
         self.weighted_samples = weighted_samples
         self.parent = parent
         self.id = id
-        self.value = np.asarray(value, dtype=np.float64)
+        self.value = np.asarray(value)
         self.is_leaf = 1
